rewrite bye and refactoring
tournament now only creates bye player if number of participants
is uneven and then sets hasBye flag. Pairing.java now has second
constructor to signal that pairing includes a bye. This eliminates
the need to check for this for all the other pairings.

Removed several commented lines that were used for print debugging