remove SimulatedPlayer.java and pull functionality into Participant
Each SimulatedTournament creates copy of every Participant in participant list, now creating a new
Participant instead of SimulatedPlayer. SimulatedTournament also keeps reference to original ParticipantList
for now, which is needed for functionality like the topThreeRanking to work. This should be rewritten.
Temporarily delete average rank functionality as it will also need to be rewritten.

Tests are currently broken.