Replace use of Map for past games with List
Since order is needed anyway for pairing to work correctly
and javafo will do all the pairing in the future there really
is no more use for the Map.

Also changed all the uses of color in TournamentDataParser to
Characters and include that information in the List that
replaced the pastResults Map.