Replace hasReceivedBye with proper TRF format
If parser sees that a player has been awarded a bye,
it will be added to the list of games the same way a
normal game would. According to the TRF format, special
values are used to save byes.

The self made pairing process is now pretty much
completely unusable.